fix: Remove redundant nil checks and refactor package naming
