+++ scrapers/common.go
@@ -1,7 +1,7 @@
 package scrapers
 
 import (
-	"github.com/flanksource/config-db/scrapers/azure/core"
+	"github.com/flanksource/config-db/scrapers/azure"
 	"time"
 
 	v1 "github.com/flanksource/config-db/api/v1"
@@ -16,7 +16,7 @@ import (
 
 // All is the scrappers registry
 var All = []v1.Scraper{
-	core.AzureScraper{},
+	azure.Scraper{},
 	aws.Scraper{},
 	aws.Scraper{},
 	aws.CostScraper{},
